fp: program index into g_prog_info_table[] should be uint16_t

+++ src/front_panel.h
@@ -19,8 +19,8 @@ int fp_build_cmd(uint8_t *buf, bool is_read, int cmd,
 		void *param, int param_size);
 bool fp_validate_cmd(uint8_t *buf, int size, int expect_param_size);
 void fp_cmd_copy(struct fp_cmd *dst, struct fp_cmd *src);
-void fp_select_program(uint8_t prog_idx);
-void fp_deselect_program(uint8_t prog_idx);
+void fp_select_program(uint16_t prog_idx);
+void fp_deselect_program(uint16_t prog_idx);
 
 
 #endif /* _FRONT_PANEL_H_ */
